Refactor Bubble functionality
The purpose of this refactor was to allow the component to keep an internal UI statae which monitors the *actual* widths of its internal textelements, rather than relying on the character width being specified through props or config (which is rarely accurate). 

The component now responds to its intermal element sizes. This means that in future iterations the component could be broken up so that it accepts text components as children. Right now however this isn't necessary. The component could also be refactored for efficiency measures; it currently reclaculates all it's props every time it is activated (isActive).

The text origin is also still needs work - which the axis is flipped the text component appears off-center, likewide when the axis are morphed.

The target of the child items also needs addressing so that an anchor/Link component is inserted in the text groups.